Now you can add multiple notifications and they will all sound
