+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.mediclock_v1">
     <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" /> 
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
      <queries>
         <package android:name="com.google.android.deskclock" />
      </queries>
