+++ pubspec.yaml
@@ -34,10 +34,8 @@ dependencies:
   flutter_local_notifications: ^6.0.0
   timezone: ^0.7.0
   intl: ^0.17.0
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.2  
+  google_sign_in: ^5.2.1  
 
 dev_dependencies:
   flutter_test:
@@ -52,6 +50,7 @@ dev_dependencies:
   flutter_launcher_icons: ^0.9.2
   url_launcher: ^6.0.0
   external_app_launcher: ^3.1.0
+  shared_preferences: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
